Update README.md
Quasar is a futuristic version of Blackjack featured in the Mass Effect video game series. Its rules are simpler to implement than Blackjack, but it has the same general idea. 
The player makes a bet and receives an initial score. At each step, the player can draw a value or stand.